% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo.R
\name{turbo}
\alias{turbo}
\title{Turbo Color Palette}
\usage{
turbo(n, start = 0, end = 1, rev = FALSE)
}
\arguments{
\item{n}{the number of colors (\emph{>= 1}) to be in the palette.}

\item{start}{the (corrected) hue in [0,1] at which the rainbow begins.}

\item{end}{the (corrected) hue in [0,1] at which the rainbow ends.}

\item{rev}{logical indicating whether the ordering of the colors should be 
reversed.}
}
\value{
A character vector \code{cv} containing n hex color codes. This can
be used either to create a user-defined color palette for subsequent graphics
by \code{\link[grDevices]{palette}(cv)}, a \code{col =} specification in 
graphics functions or in \code{par}.
}
\description{
Create a vector of n contiguous colors using the Turbo rainbow colormap.
}
\examples{
use this like e.g. grDevices::rainbow
pie(rep(1, 12), col = rainbow(12), main = "RGB/HSV")
pie(rep(1, 12), col = turbo(12), main = "turbo")
}
\references{
Mikhailkov, A. (2019). 
Google AI Blog: Turbo, An Improved Rainbow Colormap for Visualization.
\url{https://ai.googleblog.com/2019/08/turbo-improved-rainbow-colormap-for.html}.
Accessed August 26, 2020.
}
\seealso{
This function originated as (Apache 2 licensed) code at 
\url{https://gist.github.com/jlmelville/be981e2f36485d8ef9616aef60fd52ab},
in turn based on Python code at 
\url{https://gist.github.com/mikhailov-work/ee72ba4191942acecc03fe6da94fc73f}.
}
