% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizier.R
\name{embed_plotly}
\alias{embed_plotly}
\title{Embedding Plot Using the JavaScript Library Plotly}
\usage{
embed_plotly(
  coords,
  x = NULL,
  colors = NULL,
  color_scheme = NULL,
  num_colors = 15,
  alpha_scale = 1,
  limits = NULL,
  clip_limit_values = TRUE,
  title = NULL,
  show_legend = TRUE,
  cex = 1,
  text = NULL,
  tooltip = NULL,
  equal_axes = FALSE,
  pc_axes = FALSE,
  xlim = NULL,
  ylim = NULL,
  rev = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{coords}{Matrix of embedded coordinates, with as many rows as
observations, and 2 columns.}

\item{x}{Either a data frame or a column that can be used to derive a
suitable vector of colors. Ignored if \code{colors} is provided.}

\item{colors}{Vector containing colors for each coordinate.}

\item{color_scheme}{A color scheme. See 'Details'. Ignored if \code{colors}
is specified.}

\item{num_colors}{Number of distinct colors to use in the palette, when
\code{x} is a numeric vector, on the assumption that the palette is
continuous (which it probably should be). Ignored if \code{x} is not a
numeric vector. If set to \code{NULL}, it will be set to \code{length(x)}.}

\item{alpha_scale}{Scale the opacity alpha of the colors, between 0 and 1.
Useful for increasing the transparency of points, especially with large
plots with lots of overlap.}

\item{limits}{The range that the colors should map over when mapping from a
numeric vector. If not specified, then the range of \code{x}. This is useful
if there is some external absolute scale that should be used. Ignored if
\code{x} is not a numeric vector.}

\item{clip_limit_values}{If \code{TRUE} (the default) and \code{limits} is
provided, then any value that lies outside the \code{limits} is clipped to
the limiting values. If \code{tooltip = NULL} then the original values will
be preserved in the tooltip. If \code{clip_limit_values = FALSE}, then
values outside the \code{limits} will be shown in the "missing" color.}

\item{title}{Title for the plot.}

\item{show_legend}{If \code{TRUE}, display a legend. Ignored unless a
suitable categorical value is provided as \code{x} (or one can be found).}

\item{cex}{Size of the points. Ignored if \code{text} is provided.}

\item{text}{Vector of label text to display instead of a point. If the labels
are long or the data set is large, this is unlikely to be very legible, but
is occasionally useful.}

\item{tooltip}{Vector of tooltip text, to be displayed when a point is
hovered over.}

\item{equal_axes}{If \code{TRUE}, the X and Y axes are set to have the
same extents.}

\item{pc_axes}{If \code{TRUE}, the \code{coords} are replaced by the
first two (unscaled) principal components, which should have the effect of
rotating the data (with a potential reflection) so the main variance aligns
along the X-axis. Should not have any other scaling effect.}

\item{xlim}{Vector two numeric value to give the numeric extent of the
X-axis. Ignored if \code{equal_axes = TRUE} or \code{pc_axes = TRUE}.}

\item{ylim}{Vector two numeric value to give the numeric extent of the
Y-axis. Ignored if \code{equal_axes = TRUE} or \code{pc_axes = TRUE}.}

\item{rev}{logical indicating whether the ordering of the colors should be
reversed.}

\item{verbose}{If \code{TRUE}, log messages to the console, mainly when
  searching for a suitable color column in a dataframe.

More information on plotly is available at its website,
\url{https://plot.ly}.}
}
\description{
Plots the embedded coordinates, with each point colored by a specified
color, using Plotly.
}
\details{
This will open a web browser if you are using the R CLI. In RStudio, it
will put the plot in RStudio's Plots tab.

The \code{x} argument can be used to provide a suitable vector of colors
from either a data frame or vector.

If a data frame is provided, then a vector of colors will be looked for. If
it's present, it will be used as the \code{colors} argument directly.
Otherwise, a factor column will be looked for, and each level will be mapped
to a different color. Otherwise, one color will be used for each point. If
more than one column of a type is found in the data frame, the last one
encountered is used.

If a vector is provided, a similar procedure to the data frame is used when
mapping from its content to a vector of colors. Additionally, a numeric vector
can be provided, which will be linearly mapped to a color scheme.

The \code{color_scheme} parameter can be one of:
\itemize{
\item A palette function that takes an integer \code{n} and returns a vector
 of colors, e.g. \code{grDevices::rainbow}. For some other applicable
 functions, see the \code{Palettes} help page in the \code{grDevices}
 package (e.g. by running the \code{?rainbow} command).
\item A vector of colors making up a custom color scheme of your own
 devising, e.g. \code{c('red', 'green', 'blue')}. There must be at least two
 colors in the list.
\item The name of a color scheme provided by the
 \href{https://cran.r-project.org/package=paletteer}{paletteer} package, in
 the form \code{"package::palette"}. Some examples include
 \code{"dutchmasters::milkmaid"}, \code{"cartography::green.pal"},
 \code{"viridis::inferno"} and \code{"RColorBrewer::Dark2"}. If more colors
 are required than supported by the color scheme, interpolation will be used
 to create the required number of colors.
}
The default for a numeric scale is to use \code{\link[grDevices]{rainbow}}
and for factor columns to use a palette generated by 
\code{\link[Polychrome]{createPalette}}.
}
\examples{
\dontrun{
# Embed with PCA
pca_iris <- stats::prcomp(iris[, -5], retx = TRUE, rank. = 2)

# Load the plotly package
library("plotly")

# Visualize the resulting embedding, colored by iris species, using the
# rainbow color scheme
embed_plotly(pca_iris$x, iris$Species,
  color_scheme = rainbow,
  title = "iris PCA"
)

# topo.colors scheme
embed_plotly(pca_iris$x, iris$Species, color_scheme = topo.colors)

# Pass in data frame and it will use the last (in this case, only) factor
# column it finds
embed_plotly(pca_iris$x, iris)

# Don't have to see a legend if the tooltips will do
embed_plotly(pca_iris$x, iris, show_legend = FALSE)

# Custom tooltips
embed_plotly(pca_iris$x, iris, tooltip = paste("Species:", iris$Species))

# Use the "Dark2" RColorBrewer scheme
embed_plotly(pca_iris$x, iris, color_scheme = "RColorBrewer::Dark2")

# Can plot the category names instead of points, but looks bad if they're
# long (or the dataset is large)
embed_plot(pca_iris$x, iris$Species, cex = 0.5, text = iris$Species)

# Visualize numeric value (petal length) as a color
embed_plotly(pca_iris$x, iris$Petal.Length, color_scheme = "RColorBrewer::Blues")

# Can force axes to be equal size to stop cluster being distorted in one
# direction
embed_plotly(pca_iris$x, iris$Petal.Length,
  color_scheme = "RColorBrewer::Blues",
  equal_axes = TRUE
)
}
}
